feat: decouple grant applicant card from community voting component
